+++ src/types/index.ts
@@ -45,6 +45,7 @@ export * from './PushProvisioning';
 export * from './Errors';
 export * from './CustomerSheet';
 export type { Address, BillingDetails, AddressDetails } from './Common';
+export { CardBrand } from './Common';
 
 /**
  * @ignore